делит на модули
