+++ index.html
@@ -1,4 +1,5 @@
 <!DOCTYPE html>
+<script type="module" src="main.js"></script>
 <html lang="ru">
 <head>
   <meta charset="utf-8">
@@ -229,7 +230,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  
+
   <script src="js/main.js"></script>
 </body>
 </html>
